+++ preload.ts
@@ -1,19 +1,20 @@
 import { contextBridge, ipcRenderer, IpcRendererEvent } from 'electron';
+import type { FileInfo, CopyProgressData, CopyErrorData } from './main.js';
 
 contextBridge.exposeInMainWorld('api', {
   selectDirectory: (defaultPath?: string) => ipcRenderer.invoke('select-directory', defaultPath),
   scanDirectory: (dirPath: string) => ipcRenderer.invoke('scan-directory', dirPath),
-  startCopy: (files: any[], destinationDir: string) => ipcRenderer.invoke('start-copy', { files, destinationDir }),
+  startCopy: (files: FileInfo[], destinationDir: string) => ipcRenderer.invoke('start-copy', { files, destinationDir }),
   cancelCopy: () => ipcRenderer.invoke('cancel-copy'),
-  onCopyProgress: (callback: (data: any) => void) => {
-    const subscription = (_: IpcRendererEvent, data: any) => callback(data);
+  onCopyProgress: (callback: (data: CopyProgressData) => void) => {
+    const subscription = (_: IpcRendererEvent, data: CopyProgressData) => callback(data);
     ipcRenderer.on('copy-progress', subscription);
     return () => {
       ipcRenderer.removeListener('copy-progress', subscription);
     };
   },
-  onCopyError: (callback: (data: any) => void) => {
-    const subscription = (_: IpcRendererEvent, data: any) => callback(data);
+  onCopyError: (callback: (data: CopyErrorData) => void) => {
+    const subscription = (_: IpcRendererEvent, data: CopyErrorData) => callback(data);
     ipcRenderer.on('copy-error', subscription);
     return () => {
       ipcRenderer.removeListener('copy-error', subscription);